Don't bail out with a PHP Fatal error if table does not exists
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@61672 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ extensions/Farmer/MediaWikiFarmer_Wiki.php
@@ -558,6 +558,10 @@ class MediaWikiFarmer_Wiki {
 
 	protected function _deleteInterwiki() {
 		$db = $this->getDatabase();
-		$db->delete( 'interwiki', array( 'iw_prefix' => strtolower( $this->_title ) ), __METHOD__ );
+		if ( $db->tableExists ( 'interwiki' ) ) {
+			$db->delete( 'interwiki', array( 'iw_prefix' => strtolower( $this->_title ) ), __METHOD__ );
+		} else {
+			wfDebug( __METHOD__ . ": Table 'interwiki' does not exists\n" );
+		}
 	}
 }
